created options.rb file
